+++ src/main/java/me/drex/instantfeedback/block/ModBlocks.java
@@ -54,10 +54,25 @@ public class ModBlocks {
             .pushReaction(PushReaction.DESTROY));
 
     public static final Block POTTED_PALE_ROSE = register("potted_pale_rose",
-        properties -> new FlowerPotBlock(PALE_ROSE, properties), Blocks.flowerPotProperties().noOcclusion());
+        properties -> new FlowerPotBlock(PALE_ROSE, properties), Blocks.flowerPotProperties());
 
     public static final Block POTTED_CACTUS_FLOWER = register("potted_cactus_flower",
-        properties -> new FlowerPotBlock(Blocks.CACTUS_FLOWER, properties), Blocks.flowerPotProperties().noOcclusion());
+        properties -> new FlowerPotBlock(Blocks.CACTUS_FLOWER, properties), Blocks.flowerPotProperties());
+
+    public static final Block POTTED_ROSE_BUSH = register("potted_rose_bush",
+        properties -> new FlowerPotBlock(Blocks.ROSE_BUSH, properties), Blocks.flowerPotProperties());
+
+    public static final Block POTTED_PEONY = register("potted_peony",
+        properties -> new FlowerPotBlock(Blocks.PEONY, properties), Blocks.flowerPotProperties());
+
+    public static final Block POTTED_LILAC = register("potted_lilac",
+        properties -> new FlowerPotBlock(Blocks.LILAC, properties), Blocks.flowerPotProperties());
+
+    public static final Block POTTED_SUNFLOWER = register("potted_sunflower",
+        properties -> new FlowerPotBlock(Blocks.SUNFLOWER, properties), Blocks.flowerPotProperties());
+
+    public static final Block POTTED_PITCHER_PLANT = register("potted_pitcher_plant",
+        properties -> new FlowerPotBlock(Blocks.PITCHER_PLANT, properties), Blocks.flowerPotProperties());
 
     public static final Block PALE_BUSH = register(
         "pale_bush",